% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cells.R
\name{create_cells}
\alias{create_cells}
\title{create_cells}
\usage{
create_cells(
  area.sf,
  tower.dist,
  rotation.deg,
  jitter,
  small = FALSE,
  subscript,
  seed
)
}
\arguments{
\item{area.sf}{sf dataframe, focus area with geometry column (tile polygons)}

\item{tower.dist}{numeric value, distance between towers (e.g., in meters, however in general dimensionless)}

\item{rotation.deg}{numeric value, layer rotation of directional cells, sectorized, in reference to the northern direction}

\item{jitter}{numeric value, amount of jitter (e.g., in meters, however in general dimensionless)}

\item{small}{logical value, TRUE for omnidirectional cell and FALSE for directional cell}

\item{subscript}{character value, abbreviation for layer id}

\item{seed}{numeric value, seed for reproducibility}
}
\value{
A data frame with antennas of the specfic layer in the rows and the
       the following columns: point location (x, y), tower.id, cell.id, cell.kind
       (layer), intra.cell.number (in case of directional cells, three antennas per tower,
       sectorized), kind of cell (directional or omnidirectional), and rotation degree.
       This data frame corresponds to an unvalidated cell plan that can be validated
       with `create_cellplan()`.
}
\description{
Create a specific tower layer with attached directional cells
}
