% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_gen.R
\name{pop_gen}
\alias{pop_gen}
\title{pop_gen}
\usage{
pop_gen(
  tile.num,
  base.tile.size,
  city.num,
  city.size,
  city.shape = "SQUARE",
  hole.num,
  hole.size,
  hole.shape = "SQUARE",
  pop.dist.df
)
}
\arguments{
\item{tile.num}{numeric value, number of tiles in total}

\item{base.tile.size}{numeric value, size of tiles (possibly in meters, however in general without dimension)}

\item{city.num}{numeric value, number of city clusters}

\item{city.size}{numeric value, size of city cluster}

\item{city.shape}{character value, the shape of the city cluster, either "SQUARE", ...}

\item{hole.num}{numeric value, number of hole clusters}

\item{hole.size}{numeric value, size of hole clusters}

\item{hole.shape}{character value, the shape of the hole cluster, either "SQUARE", ...}

\item{pop.dist.df}{tibble with 2 columns, "type" and "expression". See Details for further instructions}
}
\value{
A list object with 7 elements, `area.params`, containing the specified parameters,
    `area.sf`, the area tibble with geometry column (tile polygons),
    `area.df`, idential to the area tibble, however, without the geometry column,
    `area.union` an sfc polygon object that contains the outter shape of the complete area
    `area.bbox` named vector with boundary coordinates of the area
    `area.raster` raster layer of the area with population values on the tile level
    `area.elevation` raster layer of the area with elevation values on the tile level
}
\description{
Generates a fully synthetic mobile phone population and geographical area.
}
\examples{
tile.num <- 100 # number of tiles
base.tile.size <- 100 # size of a single tile
city.num <- 1 # number of single cities
city.size <- 10 # size of a single city polygon
hole.num <- 1 # number of single holes
hole.size <- 1 # size of a single hole polygon
pop.dist.df <- tibble::tibble(type = c("Urban", "Hole", "Rural"),
                              expression = c("ReIns::rtpareto(shape = 0.05, scale = 10, endpoint = 200",
                                             "ReIns::rtpareto(shape = 0.7, scale = 0.1, endpoint = 20",
                                             "ReIns::rtpareto(shape = 0.5, scale = 0.1, endpoint = 100"))
area <- pop_gen(tile.num, base.tile.size,
               city.num, city.size, city.shape = "SQUARE",
               hole.num, hole.size, hole.shape = "SQUARE",
               pop.dist.df)
}
